requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
durationpy==0.10
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.106.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
posthog==5.4.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
zipp==3.23.0
